requirements.txt
python-dotenv
ruff
pytest
